backend/requirements.txt
absl-py==0.15.0
asgiref==3.5.2
astunparse==1.6.3
beautifulsoup4==4.11.1
bs4==0.0.1
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
clang==5.0
cmdstanpy==1.0.4
colorama==0.4.5
convertdate==2.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
cycler==0.11.0
Cython==0.29.30
defusedxml==0.7.1
Django==3.2.14
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
dlib==19.23.0
dukpy==0.2.3
ephem==4.1.3
filelock==3.7.1
flatbuffers==1.12
fonttools==4.34.4
gast==0.4.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.1.0
hijri-converter==2.2.4
holidays==0.14.2
idna==3.3
importlib-metadata==4.12.0
itypes==1.2.0
Jinja2==3.1.2
joblib==1.1.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
korean-lunar-calendar==0.2.1
libclang==14.0.1
LunarCalendar==0.0.9
Markdown==3.3.7
MarkupSafe==2.1.1
matplotlib==3.5.2
mlxtend==0.20.0
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==9.2.0
prophet==1.1
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.4.0
PyMeeus==0.5.11
PyPAC==0.16.0
pyparsing==3.0.9
pystan==2.19.1.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.1
regex==2022.7.9
requests==2.28.1
requests-file==1.5.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-learn==1.0.2
scipy==1.7.3
setuptools-git==1.2
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.2
tensorflow-estimator==2.8.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl==3.1.0
tldextract==3.3.1
tqdm==4.64.0
typing-extensions==3.10.0.2
ujson==5.4.0
uritemplate==4.1.1
urllib3==1.26.10
Werkzeug==2.1.2
wrapt==1.12.1
zipp==3.8.1
